llvm: Export constructor of call stacks from internal module
